+++ CHANGELOG.md
@@ -21,8 +21,7 @@
 * Updated dependencies by Nico Mexis (https://github.com/ThexXTURBOXx)
 * Added `reportOccurrenceTimeout` in CatcherOptions to prevent reporting duplicated errors.
 * Added new handler: SnackbarHandler which can be used to show snackbar message.
-* Reworked logger in Catcher. Now Catcher uses only one class to handle log printing. Default logger can be overridden with custom logger.
-* Added `logger` parameter in CatcherOptions to provide custom logger to Catcher.
+* Reworked logger in Catcher. Now Catcher uses only one class to handle log printing.
 * General logging/reporting fixes by naurmi (https://github.com/narumishi)
 * Added missing CMakeLists.txt for Linux.